move `VecPacker` and `SliceUnpacker` to the libary
