+++ bee-common/bee-common/src/packable/unpacker.rs
@@ -26,3 +26,42 @@ pub trait Unpacker: Sized {
         })
     }
 }
+
+/// A `Unpacker` backed by a `&[u8]`.
+pub struct SliceUnpacker<'u>(&'u [u8]);
+
+impl<'u> SliceUnpacker<'u> {
+    /// Create a new unpacker from a byte slice.
+    pub fn new(slice: &'u [u8]) -> Self {
+        Self(slice)
+    }
+}
+
+/// Error type to be raised when `SliceUnpacker` does not have enough bytes to unpack something.
+#[derive(Debug)]
+pub struct UnexpectedEOF {
+    /// The required number of bytes.
+    pub required: usize,
+    /// THe number of bytes the unpacker had.
+    pub had: usize,
+}
+
+impl<'u> Unpacker for SliceUnpacker<'u> {
+    type Error = UnexpectedEOF;
+
+    fn unpack_bytes(&mut self, slice: &mut [u8]) -> Result<(), Self::Error> {
+        let len = slice.len();
+
+        if self.0.len() >= len {
+            let (head, tail) = self.0.split_at(len);
+            self.0 = tail;
+            slice.copy_from_slice(head);
+            Ok(())
+        } else {
+            Err(UnexpectedEOF {
+                required: len,
+                had: self.0.len(),
+            })
+        }
+    }
+}
